Split _update_save_data() into each functions handling one version
