+++ src/elona/CMakeLists.txt
@@ -72,6 +72,7 @@ set(ELONA_SOURCES
   random.cpp
   random_event.cpp
   save.cpp
+  save_update.cpp
   set_item_info.cpp
   shop.cpp
   spec.cpp
