+++ src/elona/save_update.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "filesystem.hpp"
+
+
+
+namespace elona
+{
+
+void update_save_data(const fs::path& save_dir);
+
+}
